+++ examples/decorator.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from fastapi import FastAPI
 from pydantic import BaseModel
 
-from socketsundso import WebSocketHandlingEndpoint, on_event
+from socketsundso import WebSocketHandlingEndpoint, event
 
 # first we create the fasapi app as usual
 app = FastAPI()
@@ -30,20 +30,20 @@ class EchoMessage(BaseModel):
 @app.websocket_route("/")
 class MyWebSocketApp(WebSocketHandlingEndpoint):
 
-    # use the on_event decorator to register a handler
-    @on_event
+    # use the event decorator to register a handler
+    @event
     async def goodbye(self) -> None:
         # and of course you can reference self
         await self.websocket.send_json({"received_at": datetime.now()})
         await self.websocket.close()
 
     # your handler name can start with on_ or handle_ followed by the event name
-    @on_event()
+    @event()
     async def on_message(self, msg: str) -> typing.Dict:
         return {"msg": msg, "type": "reply"}
 
     # you can even use it for staticmethods
-    @on_event
+    @event
     @staticmethod
     async def time() -> typing.Dict:
         # but of course you won't be able to use self in here
@@ -54,13 +54,13 @@ class MyWebSocketApp(WebSocketHandlingEndpoint):
     # Other arguments include response_model.
     # For more information about all available arguments, take a look at the documentation and/or
     # code.
-    @on_event("ping", response_model=Pong)
+    @event("ping", response_model=Pong)
     async def pingpong(self) -> typing.Dict:
         return {"time": datetime.now()}
 
 
-# if you want to register a function outside of the class you have to use the on_event decorator
+# if you want to register a function outside of the class you have to use the event decorator
 # of the class. Of course you can also use all arguments with this one.
-@MyWebSocketApp.on_event("echo", response_model=EchoMessage)
+@MyWebSocketApp.event("echo", response_model=EchoMessage)
 async def outside(message: str) -> typing.Dict:
     return {"message": message}