+++ socketsundso/__init__.py
@@ -3,4 +3,4 @@
 __version__ = "0.0.3"
 
 from .endpoints import WebSocketHandlingEndpoint as WebSocketHandlingEndpoint
-from .handler import on_event as on_event
+from .handler import event as event